Restructuring of the Motus game
